+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+* Animals now have a small chance to be hostile
+  * `enhancedai:can_fight_back` entity type tag has been renamed to `enhancedai:can_be_neutral`
+  * New `enhancedai:can_be_hostile` entity type tag
+
 ## 2.5.6
 * Requires InsaneLib 1.18.3 and Forge 47.4.0
 * Wolves are no longer affected by biting mobs